refactor(policies): policies are now proper classes
