+++ lib/amqp_client.js
@@ -1,5 +1,4 @@
 'use strict';
-
 var EventEmitter = require('events').EventEmitter,
     Promise = require('bluebird'),
     debug = require('debug')('amqp10:client'),
@@ -14,11 +13,10 @@ var EventEmitter = require('events').EventEmitter,
     errors = require('./errors'),
 
     translator = require('./adapters/translate_encoder'),
-    DefaultPolicy = require('./policies/default_policy'),
+    Policy = require('./policies/policy'),
 
+    pu = require('./policies/policy_utilities'),
     u = require('./utilities');
-
-
 
 /**
  * AMQPClient is the top-level class for interacting with node-amqp10.  Instantiate this class, connect, and then send/receive
@@ -52,13 +50,14 @@ var EventEmitter = require('events').EventEmitter,
  * sent/received, and by default do a simple pass-through, leaving the encoding to/decoding from AMQP-specific types up to the library which
  * does a best-effort job.  See EventHubPolicy for a more complicated example, turning objects into UTF8-encoded buffers of JSON-strings.
  *
- * @param {DefaultPolicy} [policy]     Policy to use for connection, sessions, links, etc.  Defaults to DefaultPolicy.
+ * @param {Policy}  [policy]           Policy to use for connection, sessions, links, etc.  Defaults to DefaultPolicy.
+ * @param {Obeject} [policyOverrides]  Additional overrides for the provided policy
  * @constructor
  */
-function AMQPClient(policy) {
-  // Make a protective copy
-  this._originalPolicy = u.deepMerge(policy || DefaultPolicy);
-  this.policy = u.deepMerge(this._originalPolicy);
+function AMQPClient(policy, policyOverrides) {
+  this.policy = policy || new Policy();
+  if (!!policyOverrides) this.policy = pu.Merge(policyOverrides, this.policy);
+
   this._connection = null;
   this._session = null;
 
@@ -311,10 +310,6 @@ AMQPClient.prototype.disconnect = function() {
 AMQPClient.prototype._clearConnectionState = function(saveReconnectDetails) {
   if (!!this._connection) this._connection.removeAllListeners();
   this._connection = null;
-
-  // Copy from original to avoid any settings changes "sticking" across connections.
-  this.policy = u.deepMerge(this._originalPolicy);
-
   if (!saveReconnectDetails) {
     this._reconnect = null;
   }
