+++ lib/connection.js
@@ -136,7 +136,7 @@ var EventEmitter = require('events').EventEmitter,
  *
  * Also could be DISCARDING if an error condition triggered the CLOSE
  *
- * @param connectPolicy ConnectPolicy from a DefaultPolicy implementation
+ * @param connectPolicy ConnectPolicy from a Policy instance
  * @constructor
  */
 function Connection(connectPolicy) {
