+++ lib/policies/activemq_policy.js
@@ -1,9 +1,5 @@
 'use strict';
-
-var u = require('../utilities'),
-    DefaultPolicy = require('./default_policy');
-
-module.exports = u.deepMerge({
+var Policy = require('./policy');
+module.exports = new Policy({
   defaultSubjects: false
-}, DefaultPolicy);
-
+});
