+++ lib/policies/event_hub_policy.js
@@ -1,6 +1,2 @@
 'use strict';
-
-var u = require('../utilities'),
-    ServiceBusPolicy = require('./service_bus_policy');
-
-module.exports = u.deepMerge({}, ServiceBusPolicy);
+module.exports = require('./service_bus_policy');
