+++ lib/policies/policy_utilities.js
@@ -74,6 +74,8 @@ module.exports.Merge = merge;
 
 // Receiver links process messages N at a time, only renewing credits on ack.
 module.exports.RenewOnSettle = function(initialCredit, threshold, basePolicy) {
+  basePolicy = basePolicy || {};
+
   return merge({
     receiverLink: {
       credit: CreditPolicies.RefreshSettled(threshold),
