+++ lib/policies/qpid_java_policy.js
@@ -1,20 +1,26 @@
 'use strict';
+var Policy = require('./policy'),
+    util = require('util');
 
-var u = require('../utilities'),
-    DefaultPolicy = require('./default_policy');
+function QpidJavaPolicy() {
+  Policy.call(this, {
+    defaultSubjects: false
+  });
+}
+util.inherits(QpidJavaPolicy, Policy);
 
-module.exports = u.deepMerge({
-  defaultSubjects: false,
-  parseAddress: function(address) {
-    var result = DefaultPolicy.parseAddress(address);
+var _parseAddress = Policy.prototype.parseAddress;
+QpidJavaPolicy.prototype.parseAddress = function(address) {
+  var result = _parseAddress(address);
 
-    // accept vhosts
-    if (result.path !== '/') {
-      var pathParts = result.path.substr(1).split('/');
-      result.vhost = pathParts[0];
-      result.path = '/' + pathParts.slice(1).join('/');
-    }
-
-    return result;
+  // accept vhosts
+  if (result.path !== '/') {
+    var pathParts = result.path.substr(1).split('/');
+    result.vhost = pathParts[0];
+    result.path = '/' + pathParts.slice(1).join('/');
   }
-}, DefaultPolicy);
+
+  return result;
+};
+
+module.exports = new QpidJavaPolicy();
