+++ lib/policies/service_bus_policy.js
@@ -1,10 +1,8 @@
 'use strict';
-
 var constants = require('../constants'),
-    u = require('../utilities'),
-    DefaultPolicy = require('./default_policy');
+    Policy = require('./policy');
 
-module.exports = u.deepMerge({
+module.exports = new Policy({
   defaultSubjects: false,
   senderLink: {
     attach: {
@@ -34,6 +32,7 @@ module.exports = u.deepMerge({
       } else {
         return body; // No clue.
       }
+
       try {
         return JSON.parse(bodyStr);
       } catch (e) {
@@ -46,5 +45,4 @@ module.exports = u.deepMerge({
       forever: true
     }
   }
-}, DefaultPolicy);
-
+});
