+++ lib/utilities.js
@@ -1,5 +1,4 @@
 'use strict';
-
 var _ = require('lodash'),
     Int64 = require('node-int64'),
     uuid = require('uuid'),
@@ -7,12 +6,13 @@ var _ = require('lodash'),
     utilities = module.exports = {};
 
 // lodash aliases
-utilities.defaults = _.defaults;
+utilities.defaults = _.defaultsDeep;
 utilities.includes = _.includes;
 utilities.isObject = _.isObject;
 utilities.isNumber = _.isNumber;
 utilities.values = _.values;
 utilities.merge = _.merge;
+utilities.clone = _.cloneDeep;
 
 /**
  * Encodes given value into node-amqp-encoder format.
