+++ test/unit/test_sasl.js
@@ -4,7 +4,7 @@ var builder = require('buffer-builder'),
 
     constants = require('../../lib/constants'),
     frames = require('../../lib/frames'),
-    DefaultPolicy = require('../../lib/policies/default_policy'),
+    DefaultPolicy = require('../../lib').Policy.Default,
     pu = require('../../lib/policies/policy_utilities'),
 
     MockServer = require('./mock_amqp'),
